+++ server/Definitions/UserModification.cs
@@ -1,12 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Cryptography;
-using System.Text;
-using Napack.Common;
-using Newtonsoft.Json;
-
-namespace Napack.Server
+﻿namespace Napack.Server
 {
     public enum Operation
     {