+++ server/Modules/AdminModule.cs
@@ -36,7 +36,7 @@ namespace Napack.Server
                 return this.Response.AsJson(new
                 {
                     VersionsRecalled = 0 // TODO
-                })
+                });
             };
 
             // Deletes packages.